Move CLI helpers out of core CLI
