+++ scinoephile/cli/ocr/ocr_tesseract_cli.py
@@ -7,15 +7,15 @@ from __future__ import annotations
 from argparse import ArgumentParser
 from pathlib import Path
 
+from scinoephile.cli.helpers.io import read_image_series, write_series
 from scinoephile.common.argument_parsing import (
     get_arg_groups_by_name,
     input_file_or_dir_arg,
     output_file_arg,
 )
 from scinoephile.core import ScinoephileError
-from scinoephile.core.cli import ScinoephileCliBase, write_series
+from scinoephile.core.cli import ScinoephileCliBase
 from scinoephile.image.ocr.tesseract import ocr_image_series_with_tesseract
-from scinoephile.image.subtitles import ImageSeries
 
 __all__ = ["OcrTesseractCli"]
 
@@ -160,15 +160,7 @@ class OcrTesseractCli(ScinoephileCliBase):
             parser.error("--detect-italics may only be used with --language eng")
 
         # Read inputs
-        try:
-            image_series = ImageSeries.load(infile_path)
-        except (
-            FileNotFoundError,
-            NotADirectoryError,
-            ScinoephileError,
-            ValueError,
-        ) as exc:
-            parser.error(str(exc))
+        image_series = read_image_series(parser, infile_path)
 
         # Perform operations
         try:
