+++ test/cli/helpers/test_io.py
@@ -1,6 +1,6 @@
 #  Copyright 2017-2026 Karl T Debiec. All rights reserved. This software may be modified
 #  and distributed under the terms of the BSD license. See the LICENSE file for details.
-"""Tests of shared subtitle-series CLI input helpers."""
+"""Tests of shared CLI input and output helpers."""
 
 from __future__ import annotations
 
@@ -12,9 +12,42 @@ from unittest.mock import patch
 
 import pytest
 
+from scinoephile.cli.helpers.io import read_image_series, read_series, write_series
 from scinoephile.common.exceptions import NotAFileError
 from scinoephile.core import ScinoephileError
-from scinoephile.core.cli import read_series
+from scinoephile.core.subtitles import Series, Subtitle
+
+
+@pytest.mark.parametrize(
+    "exception",
+    [
+        FileNotFoundError("missing image subtitles"),
+        NotADirectoryError("invalid image subtitle path"),
+        ScinoephileError("invalid image subtitle series"),
+        ValueError("invalid image subtitle value"),
+    ],
+)
+def test_read_image_series_maps_file_errors_to_parser_error(
+    exception: Exception, tmp_path: Path
+):
+    """Test image subtitle read failures are reported as parser errors.
+
+    Arguments:
+        exception: exception raised while loading the image subtitle series
+        tmp_path: temporary test directory
+    """
+    parser = ArgumentParser(prog="test")
+    infile_path = tmp_path / "input.sup"
+    infile_path.write_text("", encoding="utf-8")
+    stderr = StringIO()
+
+    with patch("scinoephile.cli.helpers.io.ImageSeries.load", side_effect=exception):
+        with pytest.raises(SystemExit) as excinfo:
+            with redirect_stderr(stderr):
+                read_image_series(parser, infile_path)
+
+    assert excinfo.value.code == 2
+    assert str(exception) in stderr.getvalue()
 
 
 @pytest.mark.parametrize(
@@ -41,7 +74,7 @@ def test_read_series_maps_file_errors_to_parser_error(
     infile_path.write_text("", encoding="utf-8")
     stderr = StringIO()
 
-    with patch("scinoephile.core.cli.Series.load", side_effect=exception):
+    with patch("scinoephile.cli.helpers.io.Series.load", side_effect=exception):
         with pytest.raises(SystemExit) as excinfo:
             with redirect_stderr(stderr):
                 read_series(parser, infile_path)
@@ -66,11 +99,30 @@ def test_read_series_maps_stdin_errors_to_parser_error(exception: Exception):
     parser = ArgumentParser(prog="test")
     stderr = StringIO()
 
-    with patch("scinoephile.core.cli.stdin", StringIO("not srt")):
-        with patch("scinoephile.core.cli.Series.from_string", side_effect=exception):
+    with patch("scinoephile.cli.helpers.io.stdin", StringIO("not srt")):
+        with patch(
+            "scinoephile.cli.helpers.io.Series.from_string", side_effect=exception
+        ):
             with pytest.raises(SystemExit) as excinfo:
                 with redirect_stderr(stderr):
                     read_series(parser, "-", allow_stdin=True)
 
     assert excinfo.value.code == 2
     assert str(exception) in stderr.getvalue()
+
+
+def test_write_series_defaults_to_srt_format(tmp_path: Path):
+    """Test subtitle output helper writes SRT by default.
+
+    Arguments:
+        tmp_path: temporary test directory
+    """
+    parser = ArgumentParser(prog="test")
+    series = Series(events=[Subtitle(start=1000, end=2000, text="recognized")])
+    outfile_path = tmp_path / "recognized.ass"
+
+    write_series(parser, series, outfile_path, overwrite=True)
+
+    output_text = outfile_path.read_text(encoding="utf-8")
+    assert output_text.startswith("1\n00:00:01,000 --> 00:00:02,000\n")
+    assert "[Script Info]" not in output_text
